Move auditing to VersionControl.Audit
