+++ lib/lightning/projects/audit.ex
@@ -7,9 +7,7 @@ defmodule Lightning.Projects.Audit do
     item: "project",
     events: [
       "dataclip_retention_period_updated",
-      "history_retention_period_updated",
-      "repo_connection_created",
-      "repo_connection_removed"
+      "history_retention_period_updated"
     ]
 
   alias Ecto.Multi
@@ -47,50 +45,5 @@ defmodule Lightning.Projects.Audit do
       :changes,
       changeset.changes |> Map.filter(fn {f, _} -> f == field end)
     )
-  end
-
-  @spec repo_connection(
-          Lightning.VersionControl.ProjectRepoConnection.t(),
-          :created | :removed,
-          Lightning.Accounts.User.t()
-          | Lightning.VersionControl.ProjectRepoConnection.t()
-          | Lightning.Workflows.Trigger.t()
-        ) :: Ecto.Changeset.t()
-  def repo_connection(repo_connection, action, actor) do
-    %{
-      branch: branch,
-      config_path: config_path,
-      project_id: project_id,
-      repo: repo
-    } = repo_connection
-
-    changes =
-      %{
-        branch: branch,
-        repo: repo
-      }
-      |> then(fn connection_properties ->
-        if config_path do
-          Map.put(connection_properties, :config_path, config_path)
-        else
-          connection_properties
-        end
-      end)
-      |> then(fn connection_properties ->
-        if action == :created do
-          %{
-            after:
-              Map.put(
-                connection_properties,
-                :sync_direction,
-                repo_connection.sync_direction
-              )
-          }
-        else
-          %{before: connection_properties}
-        end
-      end)
-
-    event("repo_connection_#{action}", project_id, actor, changes)
   end
 end
